+++ CHANGELOG.md
@@ -19,6 +19,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 -   The event handler added by AddTypeScriptEventHandler should now
     work with all types of events
 
+### Added
+
+-   UpdateSupportFiles editor
+
 ## [0.25.0] - 2017-04-11
 
 [0.25.0]: https://github.com/atomist/rug-editors/compare/0.24.0...0.25.0